refactor: Extract __main__.py components into dedicated modules
Commit contains purely non-functional changes to code organization.

Extract components from __main__.py into dedicated modules
for better separation of concerns and maintainability:

New modules:
- runners/harness.py - TestRunner ABC with PytestRunner and HammettRunner
- ui/browse.py - ResultBrowser Textual TUI application
- ui/terminal.py - Terminal display utilities (spinner, status printer)
- stats.py - Statistics collection and reporting (Stat, collect_stat, print_stats, etc)
- utils/file_utils.py - File utilities (walk_source_files, copy_src_dir,
  copy_also_copy_files, setup_source_paths)

Test reorganization:
- Move all unit tests to tests/unit/ directory
- Organize tests by module: models/, mutation/, runners/, stats/, ui/, utils/, workers/
- Add tests/unit/utils/test_file_utils.py with tests for file utilities
- Add proper __init__.py files for test packages

Pre-commit fixes:
- Update exclude patterns in .pre-commit-config.yaml for tests/unit/data/
- Fix browse.py type annotations: callable -> Callable

Other changes:
- Fix TYPE_CHECKING import in code_coverage.py to use new module path
- Update test_e2e_result_snapshots.py to import walk_source_files from file_utils
