+++ .pre-commit-config.yaml
@@ -11,7 +11,7 @@ repos:
         exclude: \.rst$
     -   id: check-yaml
     -   id: debug-statements
-        exclude: tests/data/test_generation/invalid_syntax\.py
+        exclude: tests/unit/data/test_generation/invalid_syntax\.py
     -   id: end-of-file-fixer
     -   id: trailing-whitespace
 -   repo: https://github.com/astral-sh/ruff-pre-commit
@@ -25,7 +25,7 @@ repos:
     hooks:
     -   id: mypy
         args: [--config-file=mypy.ini]
-        exclude: (tests/|docs/)
+        exclude: (tests/|docs/|e2e_projects/)
         additional_dependencies:
         -    click>=8.0.0
         -    coverage>=7.3.0
