+++ tests/unit/test_generation_error_handling.py
